requirements.txt
cryptography==45.0.7
Flask[async]==3.1.1
Flask-SQLAlchemy==3.1.1
Flask_cors==5.0.0
gunicorn==23.0.0
gevent==25.9.1
numpy==2.3.3
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.1.2
python-dotenv==0.20.0
requests==2.32.5
SQLAlchemy==2.0.43
StrEnum==0.4.15
typing_extensions==4.12.2
urllib3==2.5.0
Werkzeug==3.1.3
beautifulsoup4==4.13.4
xlsxwriter==3.0.2
bleach==6.2.0
nltk==3.9.1
pandas==2.2.3
celery==5.5.1
redis==6.4.0
Flask-JWT-Extended==4.7.1
mysqlclient==2.2.7
psutil==7.0.0
PyJWT==2.10.1
Flask-PyMongo==3.0.1
openpyxl==3.1.5
Flask-Migrate==4.1.0
flake8==7.3.0
black==25.1.0
isort==6.0.1
pre-commit==4.3.0
flask-marshmallow==1.3.0
marshmallow-sqlalchemy==1.4.2
pytest==8.4.2